backend/requirements.txt
fastapi==0.118.0
uvicorn==0.37.0
python-dotenv==1.1.1
requests==2.32.5
python-magic==0.4.27
python-magic-bin==0.4.14 ; platform_system == "Windows"
